pyproject.toml
[project]
name = "codex-swarm"
version = "0.1.0"
description = "External orchestration layer for Codex supervisor/worker swarms"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Codex Swarm" }]
dependencies = [
  "click>=8.1.8",
  "orjson>=3.10.15",
  "pathspec>=0.12.1",
  "pydantic>=2.10.6",
  "pyyaml>=6.0.2",
  "rich>=13.9.4",
  "textual>=1.0.0",
]

[project.scripts]
codex-swarm = "codex_swarm.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.3",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codex_swarm"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
